requirements.txt
llama-index==0.12.30
llama-index-embeddings-huggingface==0.5.3
langchain-groq==0.3.2
langchain-core==0.3.52
sentence-transformers==4.1.0
python-dotenv==1.1.0
colorlog==6.9.0
flask[async]==3.1.0
pyyaml==6.0.2
